backend/requirements.txt
Django>=4.0,<5.0
psycopg2-binary
gunicorn
djangorestframework
django-cors-headers
djangorestframework-simplejwt
python-dotenv
drf-spectacular
pandas
requests
